requirements.txt
psycopg2
PyYAML
graphviz
